Libraries: Add pogo pin footprint variant using pads rather than vias
The complete via footprints will be easier to locate during insertion, but the pad variant requires less internal PCB volume.
The pad variant still uses two vias for locating. Pins 2 and 3 used because they are set as 3V3 and GND for the programming adapter, which you might want to be vias anyway. Of course the user can always change the pins from vias to pads as they see fit.